+++ runner/src/server/plugins/engine/index.ts
@@ -1,5 +1,5 @@
 export { RelativeUrl, FeedbackContextInfo, decode } from "./feedback";
-export { default as findByPostcode } from "./services/address-service";
+export { default as findByPostcode } from "./services/addressService";
 export { configureEnginePlugin } from "./configureEnginePlugin";
 export { redirectTo, redirectUrl } from "./helpers";
 export { default as Model } from "./model";
